hr-agent/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/LLM
google-genai==0.3.0

# Redis (optional caching)
redis==5.0.1

# Scheduler
APScheduler==3.10.4

# HTTP Client
httpx==0.26.0

# Email
aiosmtplib==3.0.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Report Generation
openpyxl==3.1.2
reportlab==4.0.8

# Templates
Jinja2==3.1.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2024.1

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
